Removed deprecated postSimplePayment
